podNetworkController: Disable LSP when VM migration target pod is ready
currently all pods have the LSP.Enabled option set to true by default.
During migration, it makes sense to handoff the network traffic from the
source to target pod as soon as the target pod is up and ready for
traffic.
The handoff is done by closing the LSP of the source port.
Doing this significantly reduces the migration downtime.

This traffic handoff optimization is currently restricted to VM with
only secondary interfaces, and layer2 topology.

Signed-off-by: Ram Lavi <ralavi@redhat.com>